issue64: Resolved issue #64. Implemented the following new tests for ddms\classes\command\NewRequest.
testRunSetsRelativeUrlToExpectedDefaultRelativeUrlIfRelativeUrlIsNotSpecified()

testRunSetsRelativeUrlToExpectedDefaultRelativeUrlIfRelativeUrlIsSpecifiedWithNoValue()

index.php is no longer the default.

Now by default, `ddms --new-request` will use an empty string as a default
`--relative-url` in order to not interfere with the root domain configuration setup
by the file template..

I.e., it is assumed if no `--relative-urls` are passed to `ddms --new-request` that
the Request should be for the root domain.

Also refactored ddms\classes\command\ConfigureAppOutput. Implemented the following
new test:

testRunConfiguresARequestForRootForRelativeUrlsWhoseValueIsABackslash()

This is related to the changes to `ddms --new-request`, now `ddms
--configure-app-output` will configure a root request properly if the
string `/` is specified as one of the `--relative-urls`.

All PhpUnit and PhpStan --level 8 tests are passing.